refactor(components): adjust import paths | hj | #3176
