--- packages/components/src/styles/global/index.ts
@@ -1,5 +1,0 @@
-export * from './sibling-tabindex-styles';
-export * from './slotted-anchor-styles';
-export * from './slotted-input-indicator-styles';
-export * from './slotted-image-styles';
-export * from './slotted-picture-image-styles';
