+++ packages/components/src/styles/index.ts
@@ -7,4 +7,3 @@ export * from './hover-media-query';
 export * from './prefers-color-scheme-dark-media-query';
 export * from './schemed-high-contrast-media-query';
 export * from './forced-colors-media-query';
-export * from './global';
